+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## Version 2.9
-* set defintion public in service container (required for Drupal 9, no functional change)
-
+* set definition public in service container (required for Drupal 9, no functional change)
+* add API NbrStudy generateids to generate study participant IDs for data only studies (see https://www.wrike.com/open.htm?id=706388658)
 
 ## Version 2.8
 * api to remove selected participation cases from study not progressed (see https://www.wrike.com/open.htm?id=753789895)